Okay.  Fix up the database table tournament so the index field (id) is a 64 bit #.   We don't want to run out of space after 2 billion tournaments.

+++ backend/models/common.rs
@@ -5,7 +5,7 @@ use chrono::Utc;
 pub type ID = i32;
 
 #[tsync::tsync]
-pub type BigId = i32;
+pub type BigId = i64;
 
 #[tsync::tsync]
 pub type UTC = DateTime<Utc>;
